first commit
first part of msc thesis: reproducing a 2017 paper on the application of a score-driven model to the problem of prediction football (soccer) outcomes based on past performance of teams.

The strength of each team is decomposed in attack-strength and defense-strength and allowed to vary with time. These strenghts are parameters in a discrete distribution from which the match-outcomes are assumed to be drawn.